requirements.txt
python-telegram-bot[socks]==20.3
httpx
nest_asyncio